@@ -1,10 +1,10 @@
 
 source 'https://rubygems.org' do
 
-  gem 'rails', '3.2.22.1'
+  gem 'rails', '4.2.10'
   gem 'mysql2', '0.3.21'
   gem 'thrift', '0.11.0'
-  gem 'activerecord', '3.2.22.1'
+  gem 'activerecord', '4.2.10'
   gem 'workflow2_types', '1.0.4'
   gem 'json', '1.8.6'
 
